backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.36,<3.0
pymysql==1.1.0
cryptography==41.0.7
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-docx>=0.8.11
pypdf==4.0.1
openpyxl>=3.1.2
pdfplumber>=0.10.3
xlrd==2.0.1
xlwt==1.3.0
openai>=1.0.0
google-generativeai>=0.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
